add dataAccess module and move data access code. Also fix default radio button
