+++ scripts/JewelryStyles.js
@@ -1,7 +1,5 @@
-import { getStyles, setOrderBuilder, setStyle} from "./database.js"
+import { getStyles, setOrderBuilder, setStyle} from "./dataAccess.js"
 
-const styles = getStyles()
-const orderBuilder = setOrderBuilder()
 
 
 document.addEventListener(
@@ -13,20 +11,23 @@ document.addEventListener(
         }
     }
     )
-
-export const JewelryStyles = () => {
-    let html = "<ul>"
     
-    // Use .map() for converting objects to <li> elements
+    export const JewelryStyles = () => {
+        let html = "<ul>"
+        
+        const styles = getStyles()
+        const orderBuilder = setOrderBuilder()
+    
+        // Use .map() for converting objects to <li> elements
     const listItems = styles.map(style => {
         if (orderBuilder.styleId === style.id) {
             return `<li>
-            <input type="radio" checked name="style" value="${style.id}" /> ${style.style}
+            <input type="radio" checked id="style" name="style" value="${style.id}" /> ${style.style}
             </li>`     
         }
         else {
             return `<li>
-            <input type="radio" name="style" value="${style.id}" /> ${style.style}
+            <input type="radio" id="style" name="style" value="${style.id}" /> ${style.style}
             </li>`
         }          
     }
