+++ scripts/JewelryType.js
@@ -1,7 +1,5 @@
-import { getTypes, setType, setOrderBuilder } from "./database.js"
+import { getTypes, setType, setOrderBuilder } from "./dataAccess.js"
 
-const type = getTypes()
-const orderBuilder = setOrderBuilder()
 
 document.addEventListener(
     "change",
@@ -11,21 +9,23 @@ document.addEventListener(
             document.dispatchEvent(new CustomEvent("stateChanged"))
         }
     }
-)
-
+    )
+    
     export const JewelryType = () => {
         let html = "<ul>"
         
+        const type = getTypes()
+        const orderBuilder = setOrderBuilder()
         // Use .map() for converting objects to <li> elements
         const listItems = type.map(type => {
             if (orderBuilder.typeId === type.id) {
                 return `<li>
-                <input type="radio" checked name="type" value="${type.id}" /> ${type.type}
+                <input type="radio" checked id="type" name="type" value="${type.id}" /> ${type.type}
                 </li>`     
             }
             else {
                 return `<li>
-                <input type="radio" name="type" value="${type.id}" /> ${type.type}
+                <input type="radio" id="type" name="type" value="${type.id}" /> ${type.type}
                 </li>`
             } 
         }
