+++ scripts/Metals.js
@@ -1,7 +1,5 @@
-import { getMetals, setMetal, setOrderBuilder } from "./database.js"
+import { getMetals, setMetal, setOrderBuilder } from "./dataAccess.js"
 
-const metals = getMetals()
-const orderBuilder = setOrderBuilder()
 
 document.addEventListener(
     "change",
@@ -11,21 +9,23 @@ document.addEventListener(
             document.dispatchEvent(new CustomEvent("stateChanged"))
         }
     }
-)
-
-export const Metals = () => {
-    let html = "<ul>"
-
+    )
+    
+    export const Metals = () => {
+        let html = "<ul>"
+        
+        const metals = getMetals()
+        const orderBuilder = setOrderBuilder()
     // This is how you have been converting objects to <li> elements
     for (const metal of metals) {
         if (orderBuilder.metalId === metal.id) {
             html += `<li>
-            <input type="radio" checked name="metal" value="${metal.id}" /> ${metal.metal}
+            <input type="radio" checked id="metal" name="metal" value="${metal.id}" /> ${metal.metal}
             </li>`     
         }
         else {
             html +=`<li>
-            <input type="radio" name="metal" value="${metal.id}" /> ${metal.metal}
+            <input type="radio" id="metal" name="metal" value="${metal.id}" /> ${metal.metal}
             </li>`
         }         
     }
