+++ scripts/database.js
@@ -5,7 +5,7 @@
     modules to get copies of the state.
 
 */
-const database = {
+export const database = {
     styles: [
         { id: 1, style: "Classic", price: 500 },
         { id: 2, style: "Modern", price: 710 },
@@ -45,40 +45,40 @@ const database = {
     }
 }
 
-export const getStyles = () => {
-    return database.styles.map(style => ({...style}))
-}
-export const getSizes = () => {
-    return database.sizes.map(size => ({...size}))
-}
-export const getMetals = () => {
-    return database.metals.map(metal => ({...metal}))
-}
-export const getOrders = () => {
-    return database.customOrders.map(order => ({...order}))
-}
-export const getTypes = () => {
-    return database.jewelryType.map(type => ({...type}))
-}
+// export const getStyles = () => {
+//     return database.styles.map(style => ({...style}))
+// }
+// export const getSizes = () => {
+//     return database.sizes.map(size => ({...size}))
+// }
+// export const getMetals = () => {
+//     return database.metals.map(metal => ({...metal}))
+// }
+// export const getOrders = () => {
+//     return database.customOrders.map(order => ({...order}))
+// }
+// export const getTypes = () => {
+//     return database.jewelryType.map(type => ({...type}))
+// }
 
 
-export const setMetal = (id) => {
-    database.orderBuilder.metalId = id
-}
+// export const setMetal = (id) => {
+//     database.orderBuilder.metalId = id
+// }
 
-export const setSize = (id) => {
-    database.orderBuilder.sizeId = id
-}
+// export const setSize = (id) => {
+//     database.orderBuilder.sizeId = id
+// }
 
-export const setStyle = (id) => {
-    database.orderBuilder.styleId = id
-}
+// export const setStyle = (id) => {
+//     database.orderBuilder.styleId = id
+// }
 
-export const setType = (id) => {
-    database.orderBuilder.typeId = id
-}
+// export const setType = (id) => {
+//     database.orderBuilder.typeId = id
+// }
 
-export const setOrderBuilder = () => database.orderBuilder
+// export const setOrderBuilder = () => database.orderBuilder
 
 export const addCustomOrder = () => {
     // Copy the current state of user choices
